% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{app_tema}
\alias{app_tema}
\title{Aplicação para criar gráficos}
\usage{
app_tema(options = list())
}
\value{
Um Shiny App
}
\description{
Aplicação para criar gráficos
}
\examples{
app_tema(list(host = "0.0.0.0", port = 1702L))
}
