% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temas.R
\name{theme_esquema}
\alias{theme_esquema}
\title{Title}
\usage{
theme_esquema(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/170,
  base_rect_size = base_size/170
)

theme_esquema(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/170,
  base_rect_size = base_size/170
)
}
\arguments{
\item{base_size}{Tamanho base da fonte}

\item{base_family}{Família base da fonte}

\item{base_line_size}{Tamanho base para elementos de linha}

\item{base_rect_size}{Tamanho base para elementos de retângulos}
}
\value{
Um objeto ggplot
}
\description{
Title
}
\examples{
library(ggplot2)
g <- ggplot(data=iris, aes(x = Sepal.Length, y = Sepal.Width)) +
  geom_point(aes(color = Species))
g + theme_esquema()
g + theme_esquema()
}
