% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temas.R
\name{theme_esquema}
\alias{theme_esquema}
\title{Title}
\usage{
theme_esquema(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/170,
  base_rect_size = base_size/170
)
}
\arguments{
\item{base_rect_size}{}
}
\value{

}
\description{
Title
}
